+++ src/CM Dashboards/Install-SRSReport.ps1
@@ -703,12 +703,18 @@ Function Install-RIReport {
         [switch]$Overwrite
     )
     Begin {
+        [bool]$RSFolderExists = $false
         Write-Debug -Message "Path [$Path], ReportServerUri [$ReportServerUri], ReportFolder [$ReportFolder], Overwrite [$Overwrite]"
     }
     Process {
         Try {
             ## Check if path is a folder
-            [bool]$IsContainer = Test-Path $Path -PathType 'Container' -ErrorAction 'SilentlyContinue'
+            If ($ReportFolder -ne '/') {
+                [string]$RsFolderParent = (Split-Path -Path $ReportFolder -Parent).Replace('\', '/')
+                [string]$RsFolderLeaf = (Split-Path -Path $ReportFolder -Leaf).Replace('\', '/')
+                [string]$GetRSFolder = Get-RsFolderContent -ReportServerUri $ReportServerUri -RsFolder $RsFolderParent | Where-Object -Property 'Name' -eq $RsFolderLeaf -ErrorAction 'SilentlyContinue'
+                $RsFolderExists = -not [string]::IsNullOrEmpty($GetRSFolder)
+            }
 
             ## Get report file paths
             [string[]]$ReportFilePaths = Get-ChildItem -Path $Path -Recurse -Filter '*.rdl' | Select-Object -ExpandProperty 'FullName' -ErrorAction 'Stop'
@@ -732,9 +738,7 @@ Function Install-RIReport {
             }
 
             ## If destination does not exists, create it.
-            If ($ReportFolder -ne '/') {
-                [string]$RsFolderParent = (Split-Path -Path $ReportFolder -Parent).Replace('\', '/')
-                [string]$RsFolderLeaf = (Split-Path -Path $ReportFolder -Leaf).Replace('\', '/')
+            If (-not $RsFolderExists) {
                 New-RsFolder -ReportServerUri $ReportServerUri -Path $RsFolderParent -Name $RsFolderLeaf
             }
 
@@ -743,7 +747,7 @@ Function Install-RIReport {
                 #  Show progress
                 Show-Progress -Status "Uploading Report [$FilePath] --> [$ReportFolder]" -Loop
                 # Upload report
-                Write-RsCatalogItem -ReportServerUri $ReportServerUri -Path $FilePath -Destination $ReportFolder -Overwrite:$OverWrite -WarningAction 'SilentlyContinue'
+                Write-RsCatalogItem -ReportServerUri $ReportServerUri -Path $FilePath -Destination $ReportFolder -Overwrite:$OverWrite #-WarningAction 'SilentlyContinue'
             }
 
             ## Save result
@@ -1131,7 +1135,7 @@ Try {
         #  Show progress
         Show-Progress -Status "Installing with SQL Extensions -->"
         #  Installing reports
-        Install-RIReport -Path $Path -ReportServerUri $ReportServerUri -ReportFolder $ReportFolder -Overwrite:$OverWrite | Out-Null
+        Install-RIReport -Path $Path -ReportServerUri $ReportServerUri -ReportFolder $ReportFolder -Overwrite:$OverWrite
         #  Set shared DataSources
         Set-RIDataSourceReference -Path $Path -ReportServerUri $ReportServerUri -DataSourceRoot $DataSourceRoot -DataSourceName $DataSourceName -FilterConnection $FilterConnection
         #  Installing helper function and granting CMDB required permissions